Remove errors for give_life() and kill()
I think the errors are unnecessary, if i call give_life() on and alive
cell it should just keep it alive and kill() on a dead cell should keep
it dead.

Alive neighbours is trickier and i might make some more mistakes where i
over increase the same cell and this will help spot those issues.
