+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "conwways-game-of-life"
+name = "conways-game-of-life"
 version = "0.1.0"
 edition = "2021"
