+++ src/lib.rs
@@ -1,0 +1,2 @@
+//! Rust impleentation of John Conway's Game of Life
+pub mod cell;
