+++ src/main.rs
@@ -1,63 +1,8 @@
+use conways_game_of_life::cell::*;
 use macroquad::prelude::*;
 
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-enum CellError {
-    AliveNeighborOverflow,
-    AliveNeighborUnderflow,
-    DeadCell,
-    AliveCell,
-}
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-enum CellState {
-    Alive,
-    Dead,
-}
-
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-struct Cell {
-    pub state: CellState,
-    pub alive_neighbors: i32,
-}
-
-impl Cell {
-    pub fn give_life(&mut self) -> Result<(), CellError> {
-        if self.state == CellState::Alive {
-            Err(CellError::AliveCell)
-        } else {
-            self.state = CellState::Alive;
-            Ok(())
-        }
-    }
-
-    pub fn kill(&mut self) -> Result<(), CellError> {
-        if self.state == CellState::Dead {
-            Err(CellError::DeadCell)
-        } else {
-            self.state = CellState::Dead;
-            Ok(())
-        }
-    }
-
-    pub fn increase_alive_neighbors(&mut self) -> Result<(), CellError> {
-        if self.alive_neighbors < 8 {
-            self.alive_neighbors = self.alive_neighbors + 1;
-            Ok(())
-        } else {
-            Err(CellError::AliveNeighborOverflow)
-        }
-    }
-
-    pub fn decrease_alive_neighbors(&mut self) -> Result<(), CellError> {
-        if self.alive_neighbors > 0 {
-            self.alive_neighbors - 1;
-            Ok(())
-        } else {
-            Err(CellError::AliveNeighborUnderflow)
-        }
-    }
-}
 #[macroquad::main("Life")]
+
 async fn main() {
     let w = screen_width() as usize;
     let h = screen_height() as usize;
@@ -71,7 +16,7 @@ async fn main() {
     for row in cells.iter_mut() {
         for cell in row.iter_mut() {
             if rand::gen_range(0, 5) == 0 {
-                *cell = CellState::Alive;
+                cell.give_life();
             }
         }
     }
@@ -101,7 +46,7 @@ async fn main() {
         // Paint cells based on the cells matrix
         for (row_index, row) in cells.iter().enumerate() {
             for (col_index, &cell) in row.iter().enumerate() {
-                if cell == CellState::Alive {
+                if cell.is_alive() {
                     let x = col_index as f32 * grid_size;
                     let y = row_index as f32 * grid_size;
                     draw_rectangle(x, y, grid_size, grid_size, BLACK);
